feature: introduce SortType and UserType enums for improved follow management
