+++ app/src/main/java/com/dudoji/android/domain/model/UserType.kt
@@ -1,0 +1,7 @@
+package com.dudoji.android.domain.model
+
+enum class UserType {
+    FOLLOWER,
+    FOLLOWING,
+    NONE
+}
