+++ app/src/main/java/com/dudoji/android/domain/repository/FollowRepository.kt
@@ -1,6 +1,8 @@
 package com.dudoji.android.domain.repository
 
+import com.dudoji.android.domain.model.SortType
 import com.dudoji.android.domain.model.User
+import com.dudoji.android.domain.model.UserType
 
 interface FollowRepository {
 
@@ -15,21 +17,4 @@ interface FollowRepository {
         sortType: SortType = SortType.NEWEST,
         keyword: String = ""
     ): List<User>
-
-    enum class SortType(val value: String) {
-        NEWEST("createdAt,desc"),
-        OLDEST("createdAt,asc"),
-        FOLLOWED_AT_NEWEST("followedAt,desc"),
-        FOLLOWED_AT_OLDEST("followedAt,asc"),
-        FOLLOWING_AT_NEWEST("followingAt,desc"),
-        FOLLOWING_AT_OLDEST("followingAt,asc"),
-        NAME_ASC("name,asc"),
-        NAME_DESC("name,desc")
-    }
-
-    enum class UserType {
-        FOLLOWER,
-        FOLLOWING,
-        NONE
-    }
 }
